Fix tool presentation dependency direction
